Removed insecure mysql_query call
storeData called mysql_query directly, passing $data to it without proper quoting. The two following lines did the same thing by using prepare/execute, which is much safer.

I've just removed the call as it seems to be unnecessary anyway and poses a security-risk, as the only call to storeData right now passes a POST-variable directly.